feat: rename intent models
